storyteller/requirements.txt
aws-xray-sdk
boto3
flask
flask-cors
flask-jwt-extended
flask-jwt-extended[asymmetric_crypto]
gunicorn
pyjwt
pytz
requests
